[Anjali] : added correct program for calculating Area of Circle

+++ gcr-codebase/programming-prerequisite/AreaOfCircle.java
@@ -2,8 +2,8 @@ import java.util.Scanner;
 public class AreaOfCircle{
     public static void main(String args[]){
     Scanner sc=new Scanner(System.in);
-    float rad=sc.nextFloat();
-    float area=3.14 * rad * rad;
+    double rad=sc.nextDouble();
+    double area=3.14 * rad * rad;
     System.out.println("Area of circle is:" + area);
     }
 }
